Simplify geometry using pixel area
Fixes Antarctica artefacts in #10. Using the `--spherical-area`
toposimplify parameter correctly preserves area, but breaks geometry
near the south pole. Using `--planar-area` solves the issue, but applies
simplification unevenly with latitude.

Projecting the data to the same CRS as the final map and then
simplifying might be the better choice down the road, since the usage of
`--planar-area` parameter will be actually correct.