Bring all parts of the hovered feature to front
A bit of an ugly hack. The feature is sliced to multiple vector tiles
when the layer is loaded, all parts need to be brought forth, otherwise
just the part of the currently hovered tile is made more visible.

There are still issues with eastern Russia and the Alaskan part of USA
at lower zoom levels.